fixed users able to edit eachother hobbies

+++ resources/assets/js/components/user/UserHobby.vue
@@ -2,17 +2,20 @@
     <div>
         <h3 class="profile__subtitle">Hobbies</h3>
         <div class="m-subheader__action">
-            <a @click="edit=!edit" v-show="!edit" tabindex="" class="btn btn-success m-btn m-btn--icon m-btn--pill m--margin-bottom-15">
+            <a @click="edit=!edit" v-show="!edit && canedit" tabindex=""
+               class="btn btn-success m-btn m-btn--icon m-btn--pill m--margin-bottom-15">
               <span>
                     <i class="la la-plus"></i>
-                    <span >Edit</span>
+                    <span>Edit</span>
               </span>
             </a>
         </div>
-        <ul class="profile__tags clearfix" >
+        <ul class="profile__tags clearfix">
             <li v-for="hobby in userhobbies"><a style="color: #34bfa3">#{{hobby.name}}</a>
-                <button v-show="edit" @click="deleteRow(hobby.id)" class="btn btn-danger m-btn m-btn--icon btn-sm m-btn--icon-only m-btn--pill"><i
-                        class="icon flaticon-delete-1"></i></button></li>
+                <button v-show="edit" @click="deleteRow(hobby.id)"
+                        class="btn btn-danger m-btn m-btn--icon btn-sm m-btn--icon-only m-btn--pill"><i
+                        class="icon flaticon-delete-1"></i></button>
+            </li>
         </ul>
         <user-hobby-form v-show="edit"
                          @close="edit=false"
@@ -26,7 +29,7 @@
 
     export default {
         name: "UserHobby",
-        props: ['userid','canedit'],
+        props: ['userid', 'canedit'],
         data() {
             return {
                 id: '',
@@ -40,21 +43,22 @@
             this.fetchData();
             this.edit = this.canedit;
         },
-        methods:{
+        methods: {
 
 
-
-            deleteRow: function(userHobbyId){
-                let vm =this;
-                axios.delete('/user/'+ this.id +'/delete/hobby', {params: {userhobbyid: userHobbyId}})
-                    .then(function (response) {
-                        let index = vm.userhobbies.findIndex(function(obj){
-                            return obj.id === userHobbyId;
-                        });
-                        if (index !== -1) {
-                            vm.userhobbies.splice(index, 1);
-                        }
-                    } )
+            deleteRow: function (userHobbyId) {
+                if (this.canedit) {
+                    let vm = this;
+                    axios.delete('/user/' + this.id + '/delete/hobby', {params: {userhobbyid: userHobbyId}})
+                        .then(function (response) {
+                            let index = vm.userhobbies.findIndex(function (obj) {
+                                return obj.id === userHobbyId;
+                            });
+                            if (index !== -1) {
+                                vm.userhobbies.splice(index, 1);
+                            }
+                        })
+                }
             },
             fetchData: function () {
                 axios.get('/user/' + this.id + '/hobby')
@@ -64,17 +68,21 @@
             },
 
             upload: function (value) {
-                let vm = this;
-                axios.post('/user/' + this.id + '/update/hobby',{data: value})
-                    .then(response => {
-                        vm.userhobbies.push({
-                            id: response.data[0].id,
-                            name: response.data[0].name,
-                        });
-                        vm.edit = false;
-                    }).catch(error => {
+                if (this.canedit) {
+
+                    let vm = this;
+                    axios.post('/user/' + this.id + '/update/hobby', {data: value})
+                        .then(response => {
+                            vm.userhobbies.push({
+                                id: response.data[0].id,
+                                name: response.data[0].name,
+                            });
+                            vm.edit = false;
+                        }).catch(error => {
                     });
+                }
             },
+
 
         }
     }
